refactor: seperate the module file
