+++ Engine/Source/Runtime/ParticleRendererRenderTask/Public/ParticleRendererRenderTask.h
@@ -6,14 +6,6 @@
 
 namespace Engine
 {
-    struct ENGINE_PARTICLERENDERERRENDERTASK_API ParticleRendererRenderInstanceTaskModule : public IModule
-    {
-        INLINE_COMPILE_TIME_TYPENAME(ParticleRendererRenderInstanceTaskModule)
-	    void Initialize() override;
-	    void Shutdown() override;
-	    bool DynamicLoadable() override;
-    };
-
     struct ENGINE_PARTICLERENDERERRENDERTASK_API ParticleRendererRenderInstanceTask : public RenderInstanceTask
     {
         INLINE_COMPILE_TIME_TYPENAME(ParticleRendererRenderInstanceTask)
